Separate general test I/O from th test command
